split popup.js into more files
