+++ esbuild.config.js
@@ -8,7 +8,7 @@ const target = 'chrome'
 
 // NOTE: index.html and manifest.json are generated via this script
 // THEY DO NOT UPDATE ON SAVE when watching (the JS does)
-process.env.NODE_ENV ||= 'production'
+process.env.NODE_ENV ||= 'localhost'
 const baseUrl = process.env.NODE_ENV === 'production' ? 'https://www.convus.org' : 'http://localhost:3009'
 const version = process.env.npm_package_version
 // Generate relevant index.html file via this hack
